rename normalization files
