+++ anomalib/utils/normalization/min_max.py
